[generator] Move public SymbolTable static methods to TypeNameUtilties static class.
