+++ tools/generator/Java.Interop.Tools.Generator.ObjectModel/Symbols/CollectionSymbol.cs
@@ -100,7 +100,7 @@ namespace MonoDroid.Generation {
 				string.Format ("var {0} = {1}.FromJniHandle ({2}, {3});",
 						opt.GetSafeIdentifier (var_name),
 						GetManagedTypeName (opt),
-						opt.GetSafeIdentifier (SymbolTable.GetNativeName (var_name)),
+						opt.GetSafeIdentifier (TypeNameUtilities.GetNativeName (var_name)),
 						owned ? "JniHandleOwnership.TransferLocalRef" : "JniHandleOwnership.DoNotTransfer"),
 			};
 		}
@@ -115,7 +115,7 @@ namespace MonoDroid.Generation {
 		{
 			return new string[] {
 				string.Format ("IntPtr {0} = {1}.ToLocalJniHandle ({2});",
-						opt.GetSafeIdentifier (SymbolTable.GetNativeName (var_name)),
+						opt.GetSafeIdentifier (TypeNameUtilities.GetNativeName (var_name)),
 						GetManagedTypeName (opt),
 						opt.GetSafeIdentifier (var_name)),
 			};
@@ -123,14 +123,14 @@ namespace MonoDroid.Generation {
 
 		public string Call (CodeGenerationOptions opt, string var_name)
 		{
-			return opt.GetSafeIdentifier (SymbolTable.GetNativeName (var_name));
+			return opt.GetSafeIdentifier (TypeNameUtilities.GetNativeName (var_name));
 		}
 
 		public string[] PostCall (CodeGenerationOptions opt, string var_name)
 		{
 			return new string[]{
 				string.Format ("JNIEnv.DeleteLocalRef ({0});",
-						opt.GetSafeIdentifier (SymbolTable.GetNativeName (var_name))),
+						opt.GetSafeIdentifier (TypeNameUtilities.GetNativeName (var_name))),
 			};
 		}
 
@@ -139,11 +139,11 @@ namespace MonoDroid.Generation {
 		#region IRequireGenericMarshal implementation
 		public string GetGenericJavaObjectTypeOverride ()
 		{
-			return SymbolTable.GetGenericJavaObjectTypeOverride (managed_name, parms != null ? parms.ToString () : null);
+			return TypeNameUtilities.GetGenericJavaObjectTypeOverride (managed_name, parms != null ? parms.ToString () : null);
 		}
 		public string ToInteroperableJavaObject (string var_name)
 		{
-			return SymbolTable.GetNativeName (var_name);
+			return TypeNameUtilities.GetNativeName (var_name);
 		}
 		#endregion
 	}
